Merged the two Spring servers into a single server
Moved all the Spring code to a single code base, now the authentication and the api is in the same project. I’m thinking that the Spring app will be served from one Apache/Nginx based site and the Angular client from another.

I’ve tidied up the client so that it always calls the spring app to get data (there’s no need for the dummy data now), added the ‘withcredentials’ to the calls to the API endpoints so that they are authorized correctly and tidied up some minor code bits